backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.1
openpyxl==3.1.5
reportbro-lib==1.1.0  # ← Esta é a correta para ReportBro
Pillow==10.4.0
Werkzeug==3.0.4
